requirements.txt
mysql-connector-python==8.0.29
pandas==1.4.2
python-dotenv==0.20.0
requests==2.27.1
tidal_unofficial==0.1.0
tqdm==4.62.3
